+++ relay-event-normalization/src/normalize/mod.rs
@@ -338,6 +338,8 @@ pub struct PerformanceScoreProfile {
     pub score_components: Vec<PerformanceScoreWeightedComponent>,
     /// See [`RuleCondition`] for all available options to specify and combine conditions.
     pub condition: Option<RuleCondition>,
+    /// The version of the profile, used to isolate changes to score calculations.
+    pub version: Option<String>,
 }
 
 /// Defines the performance configuration for the project.